+++ case.md
@@ -1,5 +1,7 @@
 - профилировала ruby-prof
+
 - медленно `Array#each` => `Array#select`
+
 - перебирает сессии по юзерам, когда можно этого избежать
 
 - потом был медленный collect_stats_from_users
