+++ spec/test.rb
@@ -1,0 +1,1 @@
+require_relative '../work'
